+++ docs/about/release-notes.rst
@@ -43,6 +43,7 @@ Features
 ~~~~~~~~
 
 * Added new string-formatting options `#3017 <https://github.com/scipp/scipp/pull/3017>`_.
+* Add ``size`` property to :class:`scipp.Variable` and :class:`scipp.DataArray`.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~